+++ java/shared/java/org/signal/libsignal/internal/Native.kt
@@ -206,6 +206,8 @@ internal object Native {
   @JvmStatic
   public external fun AuthenticatedChatConnection_clear_push_token(asyncRuntime: ObjectHandle, chat: SimpleOwner): CompletableFuture<Void?>
   @JvmStatic
+  public external fun AuthenticatedChatConnection_clear_registration_lock(asyncRuntime: ObjectHandle, chat: SimpleOwner): CompletableFuture<Void?>
+  @JvmStatic
   public external fun AuthenticatedChatConnection_connect(asyncRuntime: ObjectHandle, connectionManager: ObjectHandle, username: String, password: String, receiveStories: Boolean, languages: Array<Object>): CompletableFuture<ObjectHandle>
   @JvmStatic
   public external fun AuthenticatedChatConnection_delete_username_hash(asyncRuntime: ObjectHandle, chat: SimpleOwner): CompletableFuture<Void?>