+++ node/ts/NativeNice.ts
@@ -646,6 +646,25 @@ export async function AuthenticatedChatConnection_clear_push_token({
     )
   );
 }
+export async function AuthenticatedChatConnection_clear_registration_lock({
+  asyncContext,
+  abortSignal,
+  chat: chat,
+}: {
+  asyncContext: TokioAsyncContext;
+  abortSignal?: AbortSignal;
+  chat: Native.Wrapper<Native.AuthenticatedChatConnection>;
+}): Promise<void> {
+  return identity(
+    await asyncContext.makeCancellable(
+      abortSignal,
+      Native.AuthenticatedChatConnection_clear_registration_lock(
+        asyncContext,
+        identity(chat)
+      )
+    )
+  );
+}
 export async function AuthenticatedChatConnection_delete_username_hash({
   asyncContext,
   abortSignal,
@@ -866,6 +885,15 @@ export function TESTING_ClearPushTokenTests(): Array<GrpcTestCase<void, void>> {
     identity,
     identity
   )(Native.TESTING_ClearPushTokenTests());
+}
+
+export function TESTING_ClearRegistrationLockTests(): Array<
+  GrpcTestCase<void, void>
+> {
+  return grpcTestCaseConverter(
+    identity,
+    identity
+  )(Native.TESTING_ClearRegistrationLockTests());
 }
 
 export function TESTING_CopyBackupMediaTests(): Array<