+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 v0.97.4
 
 - AuthAccountsService.setRegistrationLock() typed API
+- AuthAccountsService.clearRegistrationLock() typed API
 - AuthAccountsService.setDiscoverableByPhoneNumber() typed API
 - UnauthBackupsService.copyBackupMedia() typed API
 - node: Add full-er service SVR2 APIs that performs necessary crypto operations inside libsignal.